moved all components into core or ui & made sidebar + header heights/widths have pixel and numeric units
