--- components/DashboardPage/constants.ts
@@ -1,2 +1,0 @@
-export const SIDEBAR_WIDTH = '150px'
-export const HEADER_HEIGHT = '65px'
