+++ components/core/Banner/index.tsx
