+++ components/core/BasePage/Head.tsx
