+++ components/core/DashboardPage/constants.ts
@@ -1,0 +1,4 @@
+export const SIDEBAR_WIDTH = 120
+export const SIDEBAR_WIDTH_PX = `${SIDEBAR_WIDTH}px`
+export const HEADER_HEIGHT = 64
+export const HEADER_HEIGHT_PX = `${HEADER_HEIGHT}px`
